+++ apps/native/src/utils/misc.ts
@@ -1,44 +1,13 @@
-import { Platform } from 'react-native';
-
 export const getCustomWallets = () => {
   const wallets = [
     {
-      id: 'rn-wallet',
-      name: 'Wallet(RN)',
-      image_url:
-        'https://github.com/reown-com/reown-docs/blob/main/static/assets/home/walletkitLogo.png?raw=true',
-      mobile_link: 'rn-web3wallet://',
-      link_mode: 'https://appkit-lab.reown.com/rn_walletkit'
-    },
-    {
-      id: 'flutter-wallet',
-      name: 'Wallet(Flutter)',
-      image_url:
-        'https://github.com/reown-com/reown-docs/blob/main/static/assets/home/walletkitLogo.png?raw=true',
-      mobile_link: 'wcflutterwallet://',
-      link_mode: 'https://appkit-lab.reown.com/flutter_walletkit'
+      id: 'phantom-wallet',
+      name: 'Phantom',
+      image_url: 'https://avatars.githubusercontent.com/u/124594793?s=200&v=4',
+      mobile_link: 'phantom://',
+      link_mode: ''
     }
   ];
-
-  if (Platform.OS === 'android') {
-    wallets.push({
-      id: 'android-wallet',
-      name: 'Wallet(Android)',
-      image_url:
-        'https://github.com/reown-com/reown-docs/blob/main/static/assets/home/walletkitLogo.png?raw=true',
-      mobile_link: 'kotlin-web3wallet://',
-      link_mode: 'https://appkit-lab.reown.com/wallet'
-    });
-  } else if (Platform.OS === 'ios') {
-    wallets.push({
-      id: 'ios-wallet',
-      name: 'Wallet(iOS)',
-      image_url:
-        'https://github.com/reown-com/reown-docs/blob/main/static/assets/home/walletkitLogo.png?raw=true',
-      mobile_link: 'walletapp://',
-      link_mode: 'https://appkit-lab.reown.com/wallet'
-    });
-  }
 
   return wallets;
 };