+++ packages/appkit/src/networks/index.ts
@@ -1,6 +1,0 @@
-// -- Networks ---------------------------------------------------------------
-export * from './solana';
-export * from './bitcoin';
-
-// -- Types ---------------------------------------------------------------
-export type { AppKitNetwork } from '@reown/appkit-common-react-native';