+++ packages/common/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @reown/appkit-common-react-native
 
+## 2.0.0-alpha.1
+
+### Patch Changes
+
+- f39727b: chore: bump alpha
+
+## 2.0.0
+
+### Major Changes
+
+- 40d26c1: BREAKING CHANGE: Prepare for AppKit v2.0.0-alpha Includes duplicate adapter validation and other API refinements.
+
+### Patch Changes
+
+- 5f71dfb: feat: phantom wallet support
+
 ## 1.2.3
 
 ### Patch Changes