+++ packages/ethers/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @reown/appkit-ethers-react-native
 
+## 2.0.0-alpha.1
+
+### Patch Changes
+
+- f39727b: chore: bump alpha
+- Updated dependencies [f39727b]
+  - @reown/appkit-scaffold-utils-react-native@2.0.0-alpha.1
+  - @reown/appkit-react-native@2.0.0-alpha.1
+  - @reown/appkit-common-react-native@2.0.0-alpha.1
+  - @reown/appkit-siwe-react-native@2.0.0-alpha.1
+
+## 2.0.0
+
+### Major Changes
+
+- 40d26c1: BREAKING CHANGE: Prepare for AppKit v2.0.0-alpha Includes duplicate adapter validation and other API refinements.
+
+### Patch Changes
+
+- 5f71dfb: feat: phantom wallet support
+- Updated dependencies [5f71dfb]
+- Updated dependencies [40d26c1]
+  - @reown/appkit-scaffold-utils-react-native@2.0.0
+  - @reown/appkit-react-native@2.0.0
+  - @reown/appkit-common-react-native@2.0.0
+  - @reown/appkit-siwe-react-native@1.2.4
+
 ## 1.2.3
 
 ### Patch Changes